[SOIN] Extrait le code de tracking de Matomo
... afin de pouvoir le réutiliser dans la visite guidée
